requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
certifi==2020.4.5.1
cloudpickle==1.3.0
cycler==0.10.0
decorator==4.4.2
entrypoints==0.3
gast==0.3.3
grpcio==1.28.1
h5py==2.10.0
imutils==0.5.3
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
jupyter-client==6.1.2
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
mock==4.0.2
numpy==1.18.1
paho-mqtt==1.5.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pygame==1.9.6
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==19.0.0
scipy==1.4.1
six==1.14.0
spyder-kernels==1.9.0
tensorboard==1.13.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
Werkzeug==1.0.1
wurlitzer==2.0.0
